requirements.txt
# Core ML and Data Processing
torch>=1.9.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Database
mysql-connector-python>=8.0.0

# API and Web
requests>=2.25.0
python-dotenv>=0.19.0

# Utilities
joblib>=1.0.0
logging
